ml_service/requirements.txt
setuptools>=68.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.4.2
aiofiles==23.2.1
sentence-transformers==2.2.2
huggingface-hub==0.19.4
transformers==4.35.2
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.1+cpu
PyPDF2==3.0.1
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.26.1
email-validator==2.1.0
PyJWT==2.10.1
# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
